requirements.txt
torch
torchvision
transformers
datasets
pandas
numpy
pillow
matplotlib>
seaborn
tqdm
tensorboard
gradio
scikit-learn
opencv-python
accelerate
